# Performance improvments to CPU Anisotropic Diffusion