```
|---------------------+--------+---------+---------|
| Function            | Type   | before  | after   |
|---------------------+--------+---------+---------|
| GradientGrayscale   | float  | 1894 ms | 1218 ms |
| GradientColorImage  | float  | 5565 ms | 3616 ms |
| CurvatureGrayscale  | float  | 2273 ms | 1548 ms |
| CurvatureColorImage | float  | 6744 ms | 4478 ms |
| GradientGrayscale   | double | 1926 ms | 1200 ms |
| GradientColorImage  | double | 5589 ms | 4038 ms |
| CurvatureGrayscale  | double | 2308 ms | 1619 ms |
| CurvatureColorImage | double | 6638 ms | 5507 ms |
|---------------------+--------+---------+---------|

```